pyproject.toml
[project]
name = "vibing-prosperity"
version = "0.1.0"
description = "Prosperity trading algorithm"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.68.0",
    "httpx>=0.27",
    "jsonpickle>=3.0.0",
    "loguru>=0.7.3",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "tyro>=1.0.11",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "pandas-stubs>=3.0.0.260204",
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
pythonpath = ["scripts"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "W"]

[tool.ruff.lint.per-file-ignores]
# Deferred import after sys.path so `scripts/` resolves.
"autoresearch.py" = ["I001"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
explicit_package_bases = true
mypy_path = ["scripts"]

[[tool.mypy.overrides]]
module = "jsonpickle.*"
ignore_missing_imports = true
